Remove non-SIMD BVH4, and keep more bool calculations in SIMD format.
